Fix target that formats TOML files
With versions >0.10, taplo switched to a scratch container that does not
contain a shell. This prevents us from using the Docker container with
Earthly, which requires a shell to run the commands in the `Earthfile`.
As a workaround, we are now simply copying the binary from the official
taplo container so that we can use ith in a container with a shell.